Eslint fixes.
