+++ src/enums/business-warning-e.ts
@@ -1,0 +1,7 @@
+export enum WarningTypesE {
+  COMPLIANCE = 'COMPLIANCE',
+  FUTURE_EFFECTIVE_AMALGAMATION = 'FUTURE_EFFECTIVE_AMALGAMATION',
+  INVOLUNTARY_DISSOLUTION = 'INVOLUNTARY_DISSOLUTION',
+  MISSING_REQUIRED_BUSINESS_INFO = 'MISSING_REQUIRED_BUSINESS_INFO',
+  NOT_IN_GOOD_STANDING = 'NOT_IN_GOOD_STANDING'
+}
